backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
sqlalchemy==2.0.23

# AI & RAG
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.22
sentence-transformers==2.2.2

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
